Add `RevitSelectFamilyExtensionTest` tests to select `Family` and `FamilyType` by name.
