+++ Directory.Build.props
@@ -1,6 +1,6 @@
 <Project>
 	<PropertyGroup>
-		<Version>1.0.0-beta.4</Version>
+		<Version>1.0.0-beta.5</Version>
     <RevitVersionTests>2024</RevitVersionTests>
     <RevitVersionCoreTests>2025</RevitVersionCoreTests>
     <RevitForceOpenClose>true</RevitForceOpenClose>
